# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service manages all backend interactions for the feed the realm clients.\n\n⚠️ NOTE: All 2xx responses are wrapped in a 'Data envelope'.\nExample: { \"data\": { \"id\": \"123\", \"name\": \"Alice\" } }",
        "title": "FeedTheRealm Core Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/world/orchestrator/players": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns player counts for all worlds.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Get all world player counts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.PlayerCountsResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/webhook/servers/update": {
            "post": {
                "description": "Webhook to update running world servers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Update running servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/players": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns player counts per zone for a world.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Get world player counts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.PlayerCountsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/zones/{zone_id}/address": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns resolved dynamically routed TCP/UDP IP and Port mappings for client connections.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Fetch running container address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "World Zone Number",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.WorldAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/zones/{zone_id}/players": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Servers report active players and average player time every 2 minutes.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Update zone player count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "World Zone Number",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Player count payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdatePlayerCountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/zones/{zone_id}/start-job": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Init an orchestrator game server container to map back to this World and Zone chunk.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Start Nomad allocation Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "World Zone Number",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to start the job in test mode, which uses a different image and resource allocation",
                        "name": "test",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acknowledge Boot",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/zones/{zone_id}/status": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the online status of a specific zone in a world.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Update zone status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "World Zone Number",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/world/orchestrator/{id}/zones/{zone_id}/stop-job": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stop local processing chunk container linked to orchestrator mapping.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "world-service"
                ],
                "summary": "Shutdown job execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "World Zone Number",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acknowledge Stop",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "dtos.PlayerCountsResponse": {
            "type": "object",
            "properties": {
                "active_players": {
                    "type": "integer"
                },
                "average_player_time": {
                    "type": "integer"
                }
            }
        },
...
        "dtos.UpdatePlayerCountRequest": {
            "type": "object",
            "properties": {
                "active_players": {
                    "type": "integer"
                },
                "average_player_time": {
                    "type": "integer"
                }
            }
        },
        "dtos.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "is_online": {
                    "type": "boolean"
                }
            }
        },
...
        "dtos.WorldAddressResponse": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
...
    }
}